@@ -162,7 +162,7 @@
     "@types/react": "^16.4.18",
     "@types/react-dom": "^16.0.9",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.0.3",
     "babel-jest": "^24.8.0",
     "babel-loader": "^8.0.4",
     "babel-plugin-css-modules-transform": "^1.1.0",
@@ -175,7 +175,7 @@
     "css-loader": "^1.0.0",
     "enzyme": "^3.10.0",
     "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^4.19.1",
+    "eslint": "^6.2.2",
     "eslint-config-airbnb": "^15.0.1",
     "eslint-config-prettier": "^2.9.0",
     "eslint-import-resolver-webpack": "^0.10.1",
